Add AbstractGenerator
